# Optimize usage of "prepare for use" in draw and dispatch commands.