Prepare for use is a command that got added due to the D3D12 driver requiring a step to transition resources to their required states, meaning the function needed to be called no matter what.

Due to how the function was written to share the loop used for validation (which does not get compiled into release games), this means that command was added and serialized into the render graph for every draw call even if the API did not require it at all.

Upon testing Nuku Warriors I found the command being called 72K times per frame on the Vulkan backend despite being a noop in that API. Furthermore, this path will also be skipped when https://github.com/godotengine/godot/pull/91769 is merged and the driver reports support for enhanced barriers.

I've not measured the performance differential but it is likely the CPU times can only improve with this change as it goes from doing something that was unnecessary to doing nothing.