@Calinou 2024-05-15
For future reference, I benchmarked this change on https://github.com/godotengine/godot-benchmarks' Bunnymark 2D benchmarks, which are notoriously CPU-bound on my machine:

Benchmark | Before | After *(this PR)*
-|-|-
`rendering/bunnymark/bunnymark_canvasitem_draw_api_5000` | 810 FPS (1.23 mspf) | 830 FPS (1.20 mspf)
`rendering/bunnymark/bunnymark_canvasitem_draw_api_10000` | 413 FPS (2.42 mspf) | 428 FPS (2.33 mspf)
`rendering/bunnymark/bunnymark_canvasitem_draw_api_20000` | 216 FPS (4.62 mspf) | 220 FPS (4.54 mspf)
||
`rendering/bunnymark/bunnymark_meshinstance2d_5000` | 527 FPS (1.89 mspf) | 532 FPS (1.87 mspf)
`rendering/bunnymark/bunnymark_meshinstance2d_10000` | 264 FPS (3.78 mspf) | 266 FPS (3.75 mspf)
`rendering/bunnymark/bunnymark_meshinstance2d_20000` | 133 FPS (7.51 mspf) | 134 FPS (7.46 mspf)
`rendering/bunnymark/bunnymark_sprite2d_5000` | 566 FPS (1.76 mspf) | 569 FPS (1.75 mspf)
`rendering/bunnymark/bunnymark_sprite2d_10000` | 278 FPS (3.59 mspf) | 285 FPS (3.50 mspf)
`rendering/bunnymark/bunnymark_sprite2d_20000` | 136 FPS (7.35 mspf) | 141 FPS (7.09 mspf)

<details open>

<summary>PC specifications</summary>

- **CPU:** Intel Core i9-13900K
- **GPU:** NVIDIA GeForce RTX 4090
- **RAM:** 64 GB (2×32 GB DDR5-5800 C30)
- **SSD:** Solidigm P44 Pro 2 TB
- **OS:** Linux (Fedora 39)
</details>